Split chutoro core into modules
